add implementation for specs2
refactor robolectric initialization to RoboTest

the content of the run() method is split into setup() and teardown()

scalatest/specs2 deps are scoped as 'provided' so the user doesn't need
to depend on both
